+++ src/main/kotlin/lotto/domain/LottoMachine.kt
@@ -1,4 +1,7 @@
-package lotto.model
+package lotto.domain
+
+import lotto.domain.model.Lotto
+import lotto.domain.model.LottoPurchaseInfo
 
 interface LottoMachine {
     fun publishLottoTickets(lottoPurchaseInfo: LottoPurchaseInfo): List<Lotto>
