+++ src/main/kotlin/lotto/domain/ManualLottoMachine.kt
@@ -1,4 +1,7 @@
-package lotto.model
+package lotto.domain
+
+import lotto.domain.model.Lotto
+import lotto.domain.model.LottoPurchaseInfo
 
 class ManualLottoMachine(private val manualNumbers: List<List<Int>>) : LottoMachine {
     override fun publishLottoTickets(lottoPurchaseInfo: LottoPurchaseInfo): List<Lotto> {
