+++ src/test/kotlin/lotto/domain/LottoMatcherTest.kt
@@ -1,5 +1,8 @@
-package lotto.model
+package lotto.domain
 
+import lotto.domain.model.Lotto
+import lotto.domain.model.LottoNumber
+import lotto.domain.model.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
