+++ src/test/kotlin/lotto/domain/model/LottoTest.kt
@@ -1,4 +1,4 @@
-package lotto.model
+package lotto.domain.model
 
 import io.kotest.matchers.shouldBe
 import org.assertj.core.api.Assertions.assertThat
